DIVISION DE LA ESTRUCTURA DE LAS FUNCIONES JS
